Add ws_edit_env_aux feature

+++ git_utils.sh
@@ -44,29 +44,23 @@
             esac
             ;;
         esac
-        if [[ -n "$del_execution" ]]; then
+        [[ -n "$del_execution" ]] && {
             for delcmd in "${del_execution[@]}"; do
                 $delcmd $branch
             done
-        fi
+        }
     done
 }
 
 ¡backupbranch() {
     local backup="$(__actual_branch)-backup"
-    if [[ $1 == -sw ]]; then
-        git checkout -b $backup
-    else
-        git branch $backup
-    fi
+    [[ $1 == -sw ]] && git checkout -b $backup || git branch $backup
     echo "$backup has been created"
 }
 
 ¡orphanbranch() {
     git checkout --orphan $1
-    if [[ $2 == -r ]]; then
-        git rm -r --cached
-    fi
+    [[ $2 == -r ]] && git rm -r --cached
     echo "$1 has been created"
 }
